Decompose settings.py into several files in settings directory
