+++ py-django-crewai/movie_chatbot/settings/base.py
@@ -1,0 +1,55 @@
+# movie_chatbot/settings/base.py
+
+import os
+from pathlib import Path # <--- Add this import
+
+# --- Define BASE_DIR here ---
+# Path(__file__) is settings/base.py
+# .parent is settings/
+# .parent.parent is movie_chatbot/
+# .parent.parent.parent is the project root (where manage.py is)
+BASE_DIR = Path(__file__).resolve().parent.parent.parent
+
+# --- Core Settings ---
+
+# SECURITY WARNING: keep the secret key used in production secret!
+SECRET_KEY = os.getenv('DJANGO_SECRET_KEY', 'django-insecure-default-dev-key-replace-in-prod')
+
+# SECURITY WARNING: don't run with debug turned on in production!
+DEBUG = os.getenv('DJANGO_DEBUG', 'False').lower() == 'true'
+
+ALLOWED_HOSTS = os.getenv('DJANGO_ALLOWED_HOSTS', '*').split(',')
+
+ROOT_URLCONF = 'movie_chatbot.urls'
+WSGI_APPLICATION = 'movie_chatbot.wsgi.application'
+
+# Default primary key field type
+DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
+
+
+# --- Internationalization ---
+# https://docs.djangoproject.com/en/4.2/topics/i18n/
+
+LANGUAGE_CODE = 'en-us'
+TIME_ZONE = 'UTC'
+USE_I18N = True
+USE_TZ = True
+
+
+# --- Password Validation ---
+# https://docs.djangoproject.com/en/4.2/ref/settings/#auth-password-validators
+
+AUTH_PASSWORD_VALIDATORS = [
+    {
+        'NAME': 'django.contrib.auth.password_validation.UserAttributeSimilarityValidator',
+    },
+    {
+        'NAME': 'django.contrib.auth.password_validation.MinimumLengthValidator',
+    },
+    {
+        'NAME': 'django.contrib.auth.password_validation.CommonPasswordValidator',
+    },
+    {
+        'NAME': 'django.contrib.auth.password_validation.NumericPasswordValidator',
+    },
+]
