+++ py-django-crewai/movie_chatbot/settings/templates.py
@@ -1,0 +1,23 @@
+# movie_chatbot/settings/templates.py
+
+import os
+from .base import BASE_DIR # Import BASE_DIR
+
+# --- Templates ---
+
+TEMPLATES = [
+    {
+        'BACKEND': 'django.template.backends.django.DjangoTemplates',
+        # Correct DIRS path using BASE_DIR
+        'DIRS': [os.path.join(BASE_DIR, 'templates')],
+        'APP_DIRS': True,
+        'OPTIONS': {
+            'context_processors': [
+                'django.template.context_processors.debug',
+                'django.template.context_processors.request',
+                'django.contrib.auth.context_processors.auth',
+                'django.contrib.messages.context_processors.messages',
+            ],
+        },
+    },
+]
